Upgrading to Laravel 6

+++ migrations/2014_10_28_175710_create_messages_table.php
@@ -15,8 +15,8 @@ class CreateMessagesTable extends Migration
     public function up()
     {
         Schema::create(Models::table('messages'), function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('thread_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('thread_id')->unsigned();
             $table->integer('user_id')->unsigned();
             $table->text('body');
             $table->timestamps();
